+++ lib/app_runner.dart
@@ -6,7 +6,6 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_native_splash/flutter_native_splash.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:google_fonts/google_fonts.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
 import 'package:eigen_engine/core/config/app_config.dart';
 import 'package:eigen_engine/core/game/game_module.dart';
@@ -56,8 +55,6 @@ Future<void> runEngineApp({
     url: config.engine.supabaseUrl,
     publishableKey: config.engine.supabasePublishableKey,
   );
-
-  GoogleFonts.config.allowRuntimeFetching = false;
 
   runApp(
     ProviderScope(